refactor: Simplify updating logic, put array in a json file and fix typings
